Switched to using floats for offsets for buttery smooth scrolling action.

+++ tenpatch/src/com/ray3k/tenpatch/TenPatchDrawable.java
@@ -305,12 +305,12 @@ public class TenPatchDrawable extends TextureRegionDrawable {
         if (!tilingX && !tilingY || !this.tiling) {
             drawToBatch(batch, texture, x + originX, y + originY, drawWidth, drawHeight, drawU, drawV, drawU2, drawV2);
         } else {
-            int offsetXadjusted = (int) (offsetX % (texX2 - texX1));
-            if (offsetXadjusted < 0) offsetXadjusted = (int) ((texX2 - texX1) + offsetXadjusted);
-            int offsetYadjusted = (int) (offsetY % (texY2 - texY1));
-            if (offsetYadjusted < 0) offsetYadjusted = (int) ((texY2 - texY1) + offsetYadjusted);
+            float offsetXadjusted = offsetX % (texX2 - texX1);
+            if (offsetXadjusted < 0) offsetXadjusted = (texX2 - texX1) + offsetXadjusted;
+            float offsetYadjusted = offsetY % (texY2 - texY1);
+            if (offsetYadjusted < 0) offsetYadjusted = (texY2 - texY1) + offsetYadjusted;
             
-            int i, j;
+            float i, j;
             
             //partial row as a result of offsetY
             if (tilingY) {
